First pass at change generation logic for missing addr:city
This is the first pass for fixing issue #3. It will detect where an addr:city tag can be added to an existing object and add the information required.

This information is not yet serialized to a .osc file but sufficient information is present within the tables to do so.

This code has not yet been tested for anything other than ways.